# specs/001-crud-empleados/contracts/empleados-openapi.yaml
openapi: 3.0.3
info:
  title: Empleados CRUD API
  version: 1.1.0
  description: >-
    API contract for employee CRUD operations under /api/v1 with server-generated employee key,
    paginated listing, and HTTP Basic authentication.
servers:
  - url: http://localhost:8080
security:
  - basicAuth: []
paths:
  /api/v1/empleados:
    get:
      summary: List employees (paginated)
      operationId: listEmpleados
      description: >-
        Returns employees in a fixed ascending order by clave_numero using
        envelope {data, pagination}. Pagination is zero-based.
      parameters:
        - name: page
          in: query
          required: false
          description: Zero-based page index. Defaults to 0 when omitted.
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          required: false
          description: Page size. Defaults to 10 when omitted. Max 100.
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
      responses:
        '200':
          description: Paginated employee list returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmpleadoPageResponse'
        '400':
          description: Invalid pagination parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    post:
      summary: Create employee
      operationId: createEmpleado
      description: Creates an employee and auto-generates `clave` as `EMP-{numero}` without leading-zero padding.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmpleadoCreateRequest'
      responses:
        '201':
          description: Employee created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmpleadoResponse'
        '400':
          description: Validation error (invalid fields or manual `clave` in payload)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: Conflict on unique key generation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /api/v1/empleados/{clave}:
    parameters:
      - name: clave
        in: path
        required: true
        schema:
          type: string
          pattern: '^EMP-[1-9][0-9]*$'
    get:
      summary: Get employee by clave
      operationId: getEmpleadoByClave
      responses:
        '200':
          description: Employee returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmpleadoResponse'
        '404':
          description: Employee not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '400':
          description: Invalid `clave` format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    put:
      summary: Update employee by clave
      operationId: updateEmpleadoByClave
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmpleadoUpdateRequest'
      responses:
        '200':
          description: Employee updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmpleadoResponse'
        '404':
          description: Employee not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '400':
          description: Invalid request or `clave` format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    delete:
      summary: Delete employee by clave
      operationId: deleteEmpleadoByClave
      responses:
        '204':
          description: Employee deleted
        '404':
          description: Employee not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '400':
          description: Invalid `clave` format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
  schemas:
    EmpleadoPageResponse:
      type: object
      required:
        - data
        - pagination
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/EmpleadoResponse'
        pagination:
          $ref: '#/components/schemas/PaginationMetadata'
    PaginationMetadata:
      type: object
      required:
        - page
        - size
        - totalElements
        - totalPages
      properties:
        page:
          type: integer
          minimum: 0
          description: Zero-based page index of the current slice.
        size:
          type: integer
          minimum: 1
          maximum: 100
        totalElements:
          type: integer
          format: int64
          minimum: 0
        totalPages:
          type: integer
          minimum: 0
    EmpleadoCreateRequest:
      type: object
      additionalProperties: false
      required:
        - nombre
        - direccion
        - telefono
      properties:
        nombre:
          type: string
          minLength: 1
          maxLength: 100
        direccion:
          type: string
          minLength: 1
          maxLength: 100
        telefono:
          type: string
          minLength: 1
          maxLength: 100
    EmpleadoUpdateRequest:
      type: object
      additionalProperties: false
      required:
        - nombre
        - direccion
        - telefono
      properties:
        nombre:
          type: string
          minLength: 1
          maxLength: 100
        direccion:
          type: string
          minLength: 1
          maxLength: 100
        telefono:
          type: string
          minLength: 1
          maxLength: 100
    EmpleadoResponse:
      type: object
      required:
        - clave
        - nombre
        - direccion
        - telefono
      properties:
        clave:
          type: string
          pattern: '^EMP-[1-9][0-9]*$'
        nombre:
          type: string
        direccion:
          type: string
        telefono:
          type: string
    ApiError:
      type: object
      required:
        - timestamp
        - status
        - error
        - message
        - path
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
          format: int32
        error:
          type: string
        message:
          type: string
        path:
          type: string